refactor, improve structure
