+++ server/routers/Pokemons/headers/addPokemon.js
@@ -5,7 +5,7 @@ module.exports = async (req, res) =>{
     const { name, description, owner, pokeAbilityName, pokeTypeName } = req.body
     const img = req.file.filename
 
-    const UserId =  req.session.user.id
+    // const UserId =  req.session.user.id
 
     try {
         const pokemonTypeId = await pokemon_type.findAll(
@@ -26,7 +26,7 @@ module.exports = async (req, res) =>{
             owner,
             pokemonAbilityId:  JSON.stringify(pokemonAbilityId[0].id),
             pokemonTypeId: JSON.stringify(pokemonTypeId[0].id) ,
-            UserId,
+            // UserId,
         })
         res.send(pokemonTypeId)
     } catch (error) {
