+++ server/routers/Pokemons/headers/deletePokemon.js
@@ -1,7 +1,7 @@
 const { Pokemon } = require('../../../db')
 
 module.exports = async (req, res) =>{
-    res.send("Se hizo el delete")
+
     const { id } = req.params
     try {
         await Pokemon.destroy({
@@ -9,6 +9,7 @@ module.exports = async (req, res) =>{
                 id
             }
         })
+    res.send("Se hizo el delete")
     } catch (error) {
         console.log(error)
     }
