+++ server/routers/index.js
@@ -3,12 +3,10 @@ const router = express.Router()
 
 const getAll = require('./Pokemons/index')
 const auth  = require('./auth/index')
-const login = require('./login')
 const admin = require('./admin')
 const register = require('./register')
 const authenticate = require('../middleware/authenticate')
 
-router.get('/login', login)
 router.get('/admin', authenticate, admin)
 router.get('/register', register)
 router.use('/pokemon', getAll)
