--- server/routers/login.js
@@ -1,6 +1,0 @@
-const path = require('path')
-
-module.exports = (req, res) =>{
-    res.render(path.resolve(__dirname, '../static/templates/login.pug'))
-
-}
