+++ server/static/templates/pokemon.pug
@@ -7,81 +7,6 @@ html
     link(rel="stylesheet", href="http://localhost:5000/static/styles/modal.css")
     title Pokemon       
 
-    //- style.
-
-    //-   *,
-    //-   *::before,
-    //-   *::after {
-    //-   margin: 0;
-    //-   padding: 0;
-    //-   box-sizing: border-box;
-    //-   font-family: "Roboto", sans-serif;
-    //-   }
-    //-   img{
-    //-     width: 100%;
-    //-   }
-    //-   body {
-    //-     background-color: #F0F8FF;
-    //-   }
-    //-   header{
-    //-     display: flex;
-    //-     justify-content: space-between;
-    //-     align-items:center;
-    //-     height: auto;
-    //-     padding: .5rem;
-    //-     width: 100%;
-    //-     background-color: steelblue;
-    //-   }
-    //-   nav{
-    //-     display: flex;
-    //-     margin-bottom: 2rem;
-    //-     width:100%;
-    //-     padding:.5rem;
-    //-     background-color: rgba(48, 115, 170, 0.63);;
-    //-     justify-content: center;
-    //-     align-items: center;
-    //-   }
-
-    //-   .button{
-    //-     padding: 16px;
-    //-     border: none;
-    //-     outline: none;
-    //-     cursor: pointer;
-    //-     font-weight: bold;
-    //-     margin-right:5px;
-    //-   }
-
-    //-   .button-logout{
-    //-     border-radius: 20px;
-    //-   }
-
-    //-   .card__container{
-    //-     display: grid;
-    //-     grid-template-columns: repeat(3, 300px);
-    //-     justify-content: space-evenly;
-    //-   }
-
-    //-   .card__content {
-    //-     border-radius: 3px;
-    //-     border: teal 2px solid;
-    //-     background-color: teal;
-    //-   }
-
-    //-   .card_body {
-    //-     display:flex; 
-    //-     padding: .5rem;
-    //-     text-align: center;
-    //-     align-items: center;
-    //-     justify-content: space-evenly;
-    //-     font-size: 1.3rem;
-    //-   }
-   
-      //- .card_body .card__info2{
-      //-   display: none;
-      //- }
-      .card__button{
-        font-size: xx-large;
-      }
   body 
     header
       h1 Pokemon-api
